# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "wlanpi-core",
...
    "version": "2.1.10"
  },
  "paths": {
...
    "/api/v1/network/interfaces/{interface}": {
      "get": {
        "tags": [
          "network"
        ],
        "summary": "Show All Interfaces",
        "description": "Returns all network interfaces.",
        "operationId": "show_all_interfaces_api_v1_network_interfaces__interface__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "interface",
            "in": "path",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Interface"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/IPInterface"
                    }
                  },
                  "title": "Response Show All Interfaces Api V1 Network Interfaces  Interface  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/network/interfaces": {
      "get": {
        "tags": [
          "network"
        ],
        "summary": "Show All Interfaces",
        "description": "Returns all network interfaces.",
        "operationId": "show_all_interfaces_api_v1_network_interfaces_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "interface",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Interface"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/IPInterface"
                    }
                  },
                  "title": "Response Show All Interfaces Api V1 Network Interfaces Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/network/ethernet/{interface}": {
      "get": {
        "tags": [
          "network"
        ],
        "summary": "Show All Ethernet Interfaces",
        "description": "Returns all ethernet interfaces.",
        "operationId": "show_all_ethernet_interfaces_api_v1_network_ethernet__interface__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "interface",
            "in": "path",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Interface"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/IPInterface"
                    }
                  },
                  "title": "Response Show All Ethernet Interfaces Api V1 Network Ethernet  Interface  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/network/ethernet/{interface}/vlan/{vlan}": {
      "get": {
        "tags": [
          "network"
        ],
        "summary": "Show All Ethernet Vlans",
        "description": "Returns all VLANS for a given ethernet interface.",
        "operationId": "show_all_ethernet_vlans_api_v1_network_ethernet__interface__vlan__vlan__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "interface",
            "in": "path",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Interface"
            }
          },
          {
            "name": "vlan",
            "in": "path",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Vlan"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/IPInterface"
                    }
                  },
                  "title": "Response Show All Ethernet Vlans Api V1 Network Ethernet  Interface  Vlan  Vlan  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "network"
        ],
        "summary": "Create Ethernet Vlan",
        "description": "Creates (or replaces) a VLAN on the given interface.",
        "operationId": "create_ethernet_vlan_api_v1_network_ethernet__interface__vlan__vlan__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "interface",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Interface"
            }
          },
          {
            "name": "vlan",
            "in": "path",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ],
              "title": "Vlan"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/IPInterfaceAddress"
                },
                "title": "Addresses"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "network"
        ],
        "summary": "Delete Ethernet Vlan",
        "description": "Removes a VLAN from the given interface.",
        "operationId": "delete_ethernet_vlan_api_v1_network_ethernet__interface__vlan__vlan__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "interface",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Interface"
            }
          },
          {
            "name": "vlan",
            "in": "path",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ],
              "title": "Vlan"
            }
          },
          {
            "name": "allow_missing",
            "in": "query",
            "required": false,
            "schema": {
              "default": false,
              "title": "Allow Missing"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/network/ethernet/{interface}/vlan": {
      "get": {
        "tags": [
          "network"
        ],
        "summary": "Show All Ethernet Vlans",
        "description": "Returns all VLANS for a given ethernet interface.",
        "operationId": "show_all_ethernet_vlans_api_v1_network_ethernet__interface__vlan_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "interface",
            "in": "path",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Interface"
            }
          },
          {
            "name": "vlan",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Vlan"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/IPInterface"
                    }
                  },
                  "title": "Response Show All Ethernet Vlans Api V1 Network Ethernet  Interface  Vlan Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/network/dhcp/leases": {
      "get": {
        "tags": [
          "network"
        ],
        "summary": "Show Dhcp Leases",
        "description": "Parse dhclient lease files under ``/var/lib/dhcp``.",
        "operationId": "show_dhcp_leases_api_v1_network_dhcp_leases_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DhcpLeasesResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DhcpLeasesResponse": {
        "properties": {
          "leases": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Leases"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "source"
        ],
        "title": "DhcpLeasesResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NetworkConfigResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "result": {
            "title": "Result"
          },
          "errors": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          }
        },
        "type": "object",
        "title": "NetworkConfigResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}